+++ app/src/main/java/dev/rustybite/rustysosho/domain/model/VerifyResponse.kt
@@ -4,7 +4,7 @@ import dev.rustybite.rustysosho.data.dto.auth.VerifiedUserResponse
 
 data class VerifyResponse(
     val accessToken: String,
-    val expiresAt: Int,
+    val expiresAt: Long,
     val refreshToken: String,
     val user: VerifiedUserResponse
 )
